Constants was collected in separate object
