+++ app/src/main/AndroidManifest.xml
@@ -12,7 +12,11 @@
         android:supportsRtl="true"
         android:theme="@style/Theme_MaterialDesign">
         <activity android:name=".view.MainActivity">
+            <intent-filter>
+                <action android:name="android.intent.action.MAIN" />
 
+                <category android:name="android.intent.category.LAUNCHER" />
+            </intent-filter>
         </activity>
 
         <meta-data
@@ -25,11 +29,7 @@
         </activity>
         <activity android:name=".view.layouts.LayoutActivity"
             android:exported="true">
-            <intent-filter>
-                <action android:name="android.intent.action.MAIN" />
 
-                <category android:name="android.intent.category.LAUNCHER" />
-            </intent-filter>
         </activity>
 
         <activity android:name=".view.navigation.BottomNavigationActivity"
