+++ app/src/main/java/ru/ds/materialdesign/repository/pictureOfTheDay/PictureOfTheDayResponseData.kt
@@ -1,4 +1,4 @@
-package ru.ds.materialdesign.repository
+package ru.ds.materialdesign.repository.pictureOfTheDay
 
 import com.google.gson.annotations.SerializedName
