+++ app/src/main/java/ru/ds/materialdesign/view/layouts/coordinator/CoordinatorFragment.kt
@@ -39,9 +39,6 @@ class CoordinatorFragment : Fragment() {
 
     }
 
-
-
-
     companion object {
         @JvmStatic
         fun newInstance() = CoordinatorFragment()
