+++ app/src/main/java/ru/ds/materialdesign/viewModel/AppState.kt
@@ -1,6 +1,6 @@
 package ru.ds.materialdesign.viewModel
 
-import ru.ds.materialdesign.repository.PictureOfTheDayResponseData
+import ru.ds.materialdesign.repository.pictureOfTheDay.PictureOfTheDayResponseData
 import ru.ds.materialdesign.repository.dto.mars.MarsPhotosServerResponseData
 import ru.ds.materialdesign.repository.epic.EarthEpicServerResponseData
