+++ app/src/main/res/layout/fragment_main.xml
@@ -5,7 +5,7 @@
     android:id="@+id/mainFragmentRoot"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:padding="10dp">
+   >
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/main"
